Add a tournaments app
The tournament app was too large to upload because of packages and app data. The app does not have full functionality to do CRUD with tournaments and events but setup was complete to do so. I used vue js after implementing the solution in angular, I wanted to try something new.